services/order-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiokafka==0.10.0
kafka-python==2.0.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.20.3
python-jose[cryptography]
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.7.7